move long dummy objects to own file
